Export of internal File Based Test Driver changes.
-- file_based_test_driver_oss: remove macro conflict; cleanup warnings
-- Internal Changes

PiperOrigin-RevId: 331222995
GitOrigin-RevId: 36f8372d56b6dbabc73cfd1dd6fe84d17a45917c

+++ base/lcs_hunt.h
@@ -57,7 +57,8 @@ class KeyOccurrences {
   // Given a key, returns the positions where this key occurs in the item
   // sequence with which the KeyOccurrences instance has been initialized.
   void occurrences(int key, const int** begin, const int** end) const {
-    if (key + 1 < first_match_.size()) {  // Is the key within the key-range?
+    // Is the key within the key-range?
+    if (key + 1 < static_cast<int>(first_match_.size())) {
       *begin = occurrences_.data() + first_match_[key];
       *end   = occurrences_.data() + first_match_[key + 1];
     } else {
